Refactor: Move Redispatch Logic from Spring Boot 3 and Spring Packages to Core
